+++ Models/Request/DocumentoCadastroRequest.cs
@@ -3,7 +3,7 @@ namespace Api_test.Models.Request
     public class DocumentoCadastroRequest
     {
         public string? Tipo { get; set; }
-        public int IdColaborador { get; set; }
+        public long IdColaborador { get; set; }
         public IFormFile? Arquivo { get; set; }
     }
 }
